requirements.txt
fastapi
uvicorn
requests
python-dotenv
slowapi
msal
python-jose[cryptography]
cachetools
